requirements.txt
python-dotenv
requests
pandas
openai
google-generativeai
tqdm
datasets
